Use already implemented logical operations tests
